Utils folder removed; rename files for histogram; and some more restructuring
