+++ callflow/__init__.py
@@ -1,3 +1,6 @@
 ## callflow.__init__.py
 
 from .datastructures.graphframe import GraphFrame
+from .utils import *
+from .timer import Timer
+from .logger import Log
