+++ callflow/pipeline/filter_networkx.py
@@ -1,8 +1,8 @@
 import numpy as np
 import networkx as nx
-from callflow.utils.logger import Log
 from ast import literal_eval as make_list
 
+from callflow.logger import Log
 
 class FilterNetworkX:
     def __init__(self, state):
