--- callflow/utils/networkx.py
@@ -1,71 +1,0 @@
-
-import networkx as nx
-from networkx.readwrite import json_graph
-
-# ------------------------------------------------------------------------------
-# networx utilities
-# ------------------------------------------------------------------------------
-# not sure if this is used anywhere
-# Also, why is this not consistent with the rest of the stlye (ie, actions)
-def dfs(graph, dataframe, limit):
-
-    def _dfs_recurse(root, level):
-        for node in root.children:
-            result = ""
-            if level < limit:
-                for i in range(0, level):
-                    result += "- "
-                node_df = dataframe.loc[
-                    (dataframe["nid"] == node.nid)
-                    & (dataframe["name"] == node.callpath[-1])
-                ]
-                inclusive_runtime = " time (inc) = " + str(node_df["time (inc)"].mean())
-                exclusive_runtime = " time = " + str(node_df["time"].mean())
-                # module = "Module = " + str(node_df['module'].unique()[0])
-                module = ""
-                result += (
-                    "Node = "
-                    + node.callpath[-1]
-                    + "["
-                    + module
-                    + ":"
-                    + str(node.nid)
-                    + "]"
-                    + inclusive_runtime
-                    + exclusive_runtime
-                )
-                log.info(result)
-                level += 1
-                dfs_recurse(node, level)
-
-    level = 0
-    for root in graph.roots:
-        print("Root = {0} [{1}]".format("Root", root._hatchet_nid))
-        _dfs_recurse(root, level)
-
-# ------------------------------------------------------------------------------
-def graphmltojson(graphfile, outfile):
-    # unused. cannot work without importing json
-    assert False
-    """
-	Converts GraphML file to json while adding communities/modularity groups
-	using python-louvain. JSON output is usable with D3 force layout.
-	"""
-    G = nx.read_graphml(graphfile)
-
-    # finds best community using louvain
-    #    partition = community_louvain.best_partition(G)
-
-    # adds partition/community number as attribute named 'modularitygroup'
-    #    for n,d in G.nodes_iter(data=True):
-    #        d['modularitygroup'] = partition[n]
-
-    node_link = json_graph.node_link_data(G)
-    json_data = json.dumps(node_link)
-
-    # Write to file
-    fo = open(outfile, "w")
-    fo.write(json_data)
-    fo.close()
-
-# ------------------------------------------------------------------------------
